HTTP, Mail URL collector: log downloaded sizes
To enable or ease troubleshooting, log the count of downloaded bytes
Just observed a case of potentially incomplete downloads (but not HTTP
errors) and it's hard to find out the actual sizes of past downloads.
Enabling debug logging is not an option either, as that logs the
complete responses (hundreds of MBs potentially).